+++ intelmq/tests/lib/test_parser_bot.py
@@ -3,6 +3,7 @@ import base64
 import datetime
 import unittest
 import unittest.mock as mock
+import warnings
 
 import intelmq.lib.bot as bot
 import intelmq.lib.test as test
@@ -73,6 +74,7 @@ class DummyParserBot(bot.ParserBot):
     """
 
     def parse_line(self, line, report):
+        warnings.warn('This is a warning test.')
         if line.startswith('#'):
             self.logger.info('Lorem ipsum dolor sit amet.')
             self.tempdata.append(line)